# contracts/openapi.yaml
components:
  schemas:
...
    LoginRequest:
      description: Login request payload.
      properties:
        email:
          description: User email.
          example: user@example.com
          type: string
        password:
          description: User password.
          example: password123
          type: string
      required:
      - email
      - password
      type: object
    LoginResponse:
      description: Login response payload.
      properties:
        access_token:
          description: Access token.
          type: string
        expires_in:
          description: Token expiration in seconds.
          format: int64
          type: integer
        refresh_token:
          description: Refresh token.
          type: string
        user:
          $ref: '#/components/schemas/UserInfo'
          description: Authenticated user info.
      required:
      - user
      - access_token
      - refresh_token
      - expires_in
      type: object
    LogoutRequest:
      description: Logout request.
      properties:
        refresh_token:
          description: The refresh token to invalidate.
          type: string
      required:
      - refresh_token
      type: object
...
    RefreshRequest:
      description: Refresh token request.
      properties:
        refresh_token:
          description: The refresh token.
          type: string
      required:
      - refresh_token
      type: object
    RegisterRequest:
      description: Registration request payload.
      properties:
        email:
          description: User email.
          type: string
        full_name:
          description: User full name.
          type: string
        password:
          description: User password.
          type: string
      required:
      - email
      - password
      - full_name
      type: object
...
    ResendVerificationRequest:
      description: Resend verification email request.
      properties:
        email:
          description: User email to resend verification to.
          type: string
      required:
      - email
      type: object
...
    UserInfo:
      description: User info.
      properties:
        full_name:
          description: Full name.
          example: John Doe
          type: string
        id:
          description: User ID.
          format: uuid
          type: string
      required:
      - id
      - full_name
      type: object
...
    VerifyEmailRequest:
      description: Email verification request.
      properties:
        token:
          description: The verification token from the email link.
          type: string
      required:
      - token
      type: object
    VerifyEmailResponse:
      description: Email verification response.
      properties:
        message:
          description: Success message.
          type: string
        verified:
          description: Whether email is now verified.
          type: boolean
      required:
      - message
      - verified
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: HTTP API layer with Axum routes and middleware
  license:
    name: ''
  title: zeltra-api
  version: 0.1.0
openapi: 3.1.0
paths:
  /auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: Login successful
        '401':
          description: Invalid credentials
      summary: POST /auth/login - Authenticate user and return tokens.
      tags:
      - Auth
  /auth/logout:
    post:
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      responses:
        '200':
          description: Logged out successfully
      summary: POST /auth/logout - Logout and invalidate refresh token.
      tags:
      - Auth
  /auth/refresh:
    post:
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Token refreshed successfully
        '401':
          description: Invalid or expired refresh token
      summary: POST /auth/refresh - Refresh access token using refresh token.
      tags:
      - Auth
  /auth/register:
    post:
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: User registered successfully
        '409':
          description: Email already exists
      summary: POST /auth/register - Register a new user.
      tags:
      - Auth
  /auth/resend-verification:
    post:
      operationId: resend_verification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
        required: true
      responses:
        '200':
          description: Verification email sent if account exists
      summary: POST /auth/resend-verification - Resend verification email.
      tags:
      - Auth
  /auth/verify-email:
    post:
      operationId: verify_email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailResponse'
          description: Email verified successfully
        '400':
          description: Invalid or expired token
      summary: POST /auth/verify-email - Verify user's email with token.
      tags:
      - Auth
...